+++ Cyberia.Api/Factories/Effects/Elements/CharacterBoostCriticalHitEffect.cs
@@ -17,6 +17,6 @@ public sealed record CharacterBoostCriticalHitEffect
 
     internal static CharacterBoostCriticalHitEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param1, parameters.Param2);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param1, (int)parameters.Param2);
     }
 }
