+++ Cyberia.Api/Factories/Effects/Elements/CharacterBoostWisdomEffect.cs
@@ -17,6 +17,6 @@ public sealed record CharacterBoostWisdomEffect
 
     internal static CharacterBoostWisdomEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param1, parameters.Param2);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param1, (int)parameters.Param2);
     }
 }
