+++ Cyberia.Api/Factories/Effects/Elements/CharacterGainRideEffect.cs
@@ -21,7 +21,7 @@ public sealed record CharacterGainRideEffect
 
     internal static CharacterGainRideEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param3, parameters.Param2, parameters.Param1 == 1);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param3, (int)parameters.Param2, parameters.Param1 == 1);
     }
 
     public RideData? GetRideData()
