+++ Cyberia.Api/Factories/Effects/Elements/CharacterGainXpFromLevelEffect.cs
@@ -18,7 +18,7 @@ public sealed record CharacterGainXpFromLevelEffect
 
     internal static CharacterGainXpFromLevelEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param1, parameters.Param2);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param1, (int)parameters.Param2);
     }
 
     public Description GetDescription()
