+++ Cyberia.Api/Factories/Effects/Elements/CharacterInventoryAddItemCheckEffect.cs
@@ -22,7 +22,7 @@ public sealed record CharacterInventoryAddItemCheckEffect
 
     internal static CharacterInventoryAddItemCheckEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param3, parameters.Param2, (GiveItemTarget)parameters.Param1);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param3, (int)parameters.Param2, (GiveItemTarget)parameters.Param1);
     }
 
     public ItemData? GetItemData()
