+++ Cyberia.Api/Factories/Effects/Elements/CharacterInventoryAddItemFromRandomDropEffect.cs
@@ -18,7 +18,7 @@ public sealed record CharacterInventoryAddItemFromRandomDropEffect
 
     internal static CharacterInventoryAddItemFromRandomDropEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param1, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param1, (int)parameters.Param3);
     }
 
     public Description GetDescription()
