+++ Cyberia.Api/Factories/Effects/Elements/CharacterInventoryRemoveItemAroundEffect.cs
@@ -19,7 +19,7 @@ public sealed record CharacterInventoryRemoveItemAroundEffect
 
     internal static CharacterInventoryRemoveItemAroundEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param3, parameters.Param2);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param3, (int)parameters.Param2);
     }
 
     public ItemData? GetItemData()
