+++ Cyberia.Api/Factories/Effects/Elements/CharacterLearnEmoteEffect.cs
@@ -17,7 +17,7 @@ public sealed record CharacterLearnEmoteEffect
 
     internal static CharacterLearnEmoteEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param3);
     }
 
     public EmoteData? GetEmoteData()
