+++ Cyberia.Api/Factories/Effects/Elements/CharacterMovementPointsWinEffect.cs
@@ -15,6 +15,6 @@ public sealed record CharacterMovementPointsWinEffect
 
     internal static CharacterMovementPointsWinEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param1, parameters.Param2);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param1, (int)parameters.Param2);
     }
 }
