+++ Cyberia.Api/Factories/Effects/Elements/CharacterPunishmentEffect.cs
@@ -21,7 +21,7 @@ public sealed record CharacterPunishmentEffect
 
     internal static CharacterPunishmentEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param1, parameters.Param2, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param1, (int)parameters.Param2, (int)parameters.Param3);
     }
 
     public EffectData? GetBoostEffectData()
