+++ Cyberia.Api/Factories/Effects/Elements/CharacterTransform2Effect.cs
@@ -17,7 +17,7 @@ public sealed record CharacterTransform2Effect
 
     internal static CharacterTransform2Effect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param1);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param1);
     }
 
     public MonsterData? GetMonsterData()
