+++ Cyberia.Api/Factories/Effects/Elements/FightUnsetStateEffect.cs
@@ -15,6 +15,6 @@ public sealed record FightUnsetStateEffect
 
     internal static FightUnsetStateEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param3);
     }
 }
