+++ Cyberia.Api/Factories/Effects/Elements/GotoMapEffect.cs
@@ -19,7 +19,7 @@ public sealed record GotoMapEffect
 
     internal static GotoMapEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param2, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param2, (int)parameters.Param3);
     }
 
     public MapData? GetMapData()
