+++ Cyberia.Api/Factories/Effects/Elements/ItemLivingCategoryEffect.cs
@@ -17,7 +17,7 @@ public sealed record ItemLivingCategoryEffect
 
     internal static ItemLivingCategoryEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param3);
     }
 
     public ItemTypeData? GetItemTypeData()
