+++ Cyberia.Api/Factories/Effects/Elements/ItemLivingMaxSkinEffect.cs
@@ -16,7 +16,7 @@ public sealed record ItemLivingMaxSkinEffect
 
     internal static ItemLivingMaxSkinEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param3);
     }
 
     public Description GetDescription()
