+++ Cyberia.Api/Factories/Effects/Elements/PetSetPowerBoostEffect.cs
@@ -19,7 +19,7 @@ public sealed record PetSetPowerBoostEffect
 
     internal static PetSetPowerBoostEffect Create(int effectId, EffectParameters parameters, int duration, int probability, CriteriaCollection criteria, EffectArea effectArea)
     {
-        return new(effectId, duration, probability, criteria, effectArea, parameters.Param2, parameters.Param3);
+        return new(effectId, duration, probability, criteria, effectArea, (int)parameters.Param2, (int)parameters.Param3);
     }
 
     public ItemData? GetItemData()
